# openapi/answer-engine.yaml
openapi: 3.1.0
info:
  title: Answer Engine OSS API
  version: 1.1.0
  summary: Local-first memory, retrieval, and grounded answers for AI agents.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://127.0.0.1:5050
    description: Default local server
security:
  - apiKey: []
  - bearerKey: []
...
paths:
...
  /api/v1/content/import/preview:
    post:
      tags: [Content]
      summary: Validate an import without writing it
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/content/import:
    post:
      tags: [Content]
      summary: Import or update local content
      description: Reusing a source identifier updates the matching item.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportRequest' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/first-imports:
    post:
      tags: [Workflows]
      summary: Register a metadata-only supported agent-history discovery
      description: Registers paths, file statistics, fingerprints, privacy posture, and exclusions. This endpoint never imports content.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FirstImportDiscovery' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /api/v1/first-imports/{sessionId}:
    parameters:
      - { $ref: '#/components/parameters/FirstImportSessionId' }
    get:
      tags: [Workflows]
      summary: Read a first-import session and its source-backed item manifest
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/first-imports/{sessionId}/approve:
    parameters:
      - { $ref: '#/components/parameters/FirstImportSessionId' }
    post:
      tags: [Workflows]
      summary: Explicitly approve a subset of discovered history sources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [sourceIds]
              properties:
                sourceIds:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  items: { $ref: '#/components/schemas/FirstImportSourceId' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/first-imports/{sessionId}/start:
    parameters:
      - { $ref: '#/components/parameters/FirstImportSessionId' }
    post:
      tags: [Workflows]
      summary: Start an explicitly approved first import
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/first-imports/{sessionId}/events:
    parameters:
      - { $ref: '#/components/parameters/FirstImportSessionId' }
    post:
      tags: [Workflows]
      summary: Record a safe per-history import outcome
      description: Imported outcomes are accepted only when all referenced chat rows have non-empty summaries and a matching raw-archive manifest path.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FirstImportEvent' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/first-imports/{sessionId}/cancel:
    parameters:
      - { $ref: '#/components/parameters/FirstImportSessionId' }
    post:
      tags: [Workflows]
      summary: Request safe cancellation at the next resumable boundary
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/first-imports/{sessionId}/retry:
    parameters:
      - { $ref: '#/components/parameters/FirstImportSessionId' }
    post:
      tags: [Workflows]
      summary: Reset failed or interrupted selected items for retry
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/first-imports/{sessionId}/complete:
    parameters:
      - { $ref: '#/components/parameters/FirstImportSessionId' }
    post:
      tags: [Workflows]
      summary: Finalize only after every discovered item has an outcome
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/content:
    get:
      tags: [Content]
      summary: List, filter, sort, and paginate content
      description: Returns lightweight rows with source, status, and assigned tags. Stable cursors follow the selected sort field and direction.
      parameters:
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { in: query, name: cursor, schema: { type: string } }
        - { in: query, name: libraryId, schema: { type: string } }
        - { in: query, name: librarySlug, schema: { type: string } }
        - { in: query, name: search, schema: { type: string, maxLength: 500 } }
        - { in: query, name: contentType, schema: { $ref: '#/components/schemas/ContentType' } }
        - { in: query, name: contentTypes, description: Comma-separated content types., schema: { type: string } }
        - { in: query, name: source, schema: { type: string, maxLength: 120 } }
        - { in: query, name: sources, description: Comma-separated neutral source identifiers such as claude-code, codex, and cowork., schema: { type: string } }
        - { in: query, name: tag, schema: { type: string, maxLength: 120 } }
        - { in: query, name: tags, description: Comma-separated tag slugs., schema: { type: string } }
        - { in: query, name: status, schema: { enum: [active, archived] } }
        - { in: query, name: dateFrom, schema: { type: string, format: date-time } }
        - { in: query, name: dateTo, schema: { type: string, format: date-time } }
        - { in: query, name: sortBy, schema: { enum: [createdAt, title, contentType, source, status], default: createdAt } }
        - { in: query, name: sortDirection, schema: { enum: [asc, desc], default: desc } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/content/raw-archive-references:
    get:
      tags: [Content]
      summary: List raw archive manifests still referenced by content
      description: Returns tenant-scoped manifest paths, including references held by soft-deleted content, so local retention can preserve recoverable sources.
      responses:
        '200':
          description: Referenced manifest paths.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/SuccessEnvelope' }
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RawArchiveReferences' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/content/{id}:
    parameters:
      - { $ref: '#/components/parameters/ContentId' }
    get:
      tags: [Content]
      summary: Get a content item
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Content]
      summary: Soft-delete a content item
      responses:
        '204': { description: Content deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/content/{id}/lineage:
    parameters:
      - { $ref: '#/components/parameters/ContentId' }
    get:
      tags: [Content]
      summary: Inspect source provenance and artifact history
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/tags/{tagId}/content:
    parameters:
      - { $ref: '#/components/parameters/TagId' }
    post:
      tags: [Organization]
      summary: Assign a tag to content in bulk
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagAssignment' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
    delete:
      tags: [Organization]
      summary: Remove a tag from content in bulk
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagAssignment' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    bearerKey:
      type: http
      scheme: bearer
  parameters:
    ContentId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
    TagId:
      in: path
      name: tagId
      required: true
      schema: { type: string, format: uuid }
...
    FirstImportSessionId:
      in: path
      name: sessionId
      required: true
      schema: { type: string, format: uuid }
  responses:
    Success:
      description: Request completed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    ValidationError:
      description: Input validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: A valid local API key is required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: The requested content was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: The requested mutation conflicts with protected or current resource state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    FirstImportSourceId:
      type: string
      enum: [claude-code, codex, cowork]
    FirstImportDiscovery:
      type: object
      additionalProperties: false
      required: [manifestPath, sources]
      properties:
        manifestPath: { type: string, minLength: 1 }
        sources:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: object
            additionalProperties: false
            required: [sourceId, label, paths, estimatedCount, estimatedBytes, privacyPosture, exclusions, availability, availabilityNote, items]
            properties:
              sourceId: { $ref: '#/components/schemas/FirstImportSourceId' }
              label: { type: string, minLength: 1 }
              paths: { type: array, items: { type: string, minLength: 1 } }
              estimatedCount: { type: integer, minimum: 0 }
              estimatedBytes: { type: integer, minimum: 0 }
              privacyPosture: { type: string, minLength: 1 }
              exclusions: { type: array, items: { type: string } }
              availability: { enum: [available, not_found, unsupported_platform, unavailable] }
              availabilityNote: { type: string, minLength: 1 }
              items:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  required: [fingerprint, sourcePath, byteSize, modifiedAt]
                  properties:
                    fingerprint: { type: string, pattern: '^[a-f0-9]{64}$' }
                    sourcePath: { type: string, minLength: 1 }
                    byteSize: { type: integer, minimum: 0 }
                    modifiedAt: { type: string, format: date-time }
    FirstImportEvent:
      type: object
      additionalProperties: false
      required: [sourceId, fingerprint, outcome]
      properties:
        sourceId: { $ref: '#/components/schemas/FirstImportSourceId' }
        fingerprint: { type: string, pattern: '^[a-f0-9]{64}$' }
        outcome: { enum: [imported, duplicate, failed, skipped] }
        contentIds: { type: array, items: { type: string, format: uuid } }
        archiveManifestPath: { type: string, minLength: 1 }
        errorCode: { type: string, pattern: '^[A-Z0-9_]{2,80}$' }
        recoveryAction: { type: string, minLength: 1 }
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
        meta: { type: object, additionalProperties: true }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    RawArchiveReferences:
      type: object
      additionalProperties: false
      required: [manifestPaths]
      properties:
        manifestPaths:
          type: array
          items: { type: string }
          uniqueItems: true
...
    TagAssignment:
      type: object
      additionalProperties: false
      required: [contentIds]
      properties:
        contentIds:
          type: array
          minItems: 1
          maxItems: 500
          items: { type: string, format: uuid }
...
    ImportRequest:
      type: object
      required: [items]
      properties:
        libraryId: { type: string }
        librarySlug: { type: string }
        items:
          type: array
          minItems: 1
          maxItems: 500
          items:
            type: object
            required: [title]
            properties:
              title: { type: string, maxLength: 500 }
              content: { type: string, maxLength: 5242880 }
              contentType: { $ref: '#/components/schemas/ContentType' }
              sourceIdentifier: { type: string, maxLength: 512 }
              source: { type: string, maxLength: 120 }
              sourceData: { type: object, additionalProperties: true }
              metadata: { type: object, additionalProperties: true }
              rawArchiveManifest: { type: object, additionalProperties: true }
              sourceAgentId: { type: string }
              conversationId: { type: string }
              turnIndex: { type: integer, minimum: 0 }
              turnRole:
                enum: [user, assistant, system, tool, developer, other]
              turnTimestamp: { type: string, format: date-time }
...
    ContentType:
      enum: [call, document, ticket, domain, chat, page]